Update routes/db functions to work for any table
